I refactored the fragments to use viewmodelfactory provider functions, so that when testing we can just subclass the fragment and override that function.
I have made many classes and methods for this because I didn't want to add too many default parameters to the methods and the classes constructors. If we didn't mind that, we could just create the universal case and instantiate with the parameters we need in each case. I reckon in big apps with many different viewmodels we should not do that, so no point in getting bad habits ne!